pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = "HW-08"
authors = ["Yulia  <nayanzina@gmail.com>"]

readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
alembic = "^1.15.1"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.39"
pydantic-settings = "^2.8.1"
greenlet = "^3.1.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
